chore: move cache pickle method
